LPD-69638 follow-up: move logic
