+++ modules/dxp/apps/ai-hub/ai-hub-rest-impl/src/main/java/com/liferay/ai/hub/rest/internal/resource/v1_0/AgentInstanceResourceImpl.java
@@ -5,9 +5,27 @@
 
 package com.liferay.ai.hub.rest.internal.resource.v1_0;
 
+import com.liferay.ai.hub.rest.dto.v1_0.AgentInstance;
+import com.liferay.ai.hub.rest.internal.resource.v1_0.util.WorkflowContextUtil;
 import com.liferay.ai.hub.rest.resource.v1_0.AgentInstanceResource;
+import com.liferay.ai.hub.rest.resource.v1_0.util.SseUtil;
+import com.liferay.ai.hub.util.AccountEntryUtil;
+import com.liferay.portal.kernel.feature.flag.FeatureFlagManagerUtil;
+import com.liferay.portal.kernel.workflow.WorkflowDefinition;
+import com.liferay.portal.kernel.workflow.WorkflowInstance;
+import com.liferay.portal.kernel.workflow.WorkflowInstanceManager;
+import com.liferay.portal.workflow.manager.WorkflowDefinitionManager;
+
+import jakarta.ws.rs.core.Context;
+import jakarta.ws.rs.sse.Sse;
+import jakarta.ws.rs.sse.SseEventSink;
+
+import java.io.Serializable;
+
+import java.util.Map;
 
 import org.osgi.service.component.annotations.Component;
+import org.osgi.service.component.annotations.Reference;
 import org.osgi.service.component.annotations.ServiceScope;
 
 /**
@@ -18,4 +36,71 @@ import org.osgi.service.component.annotations.ServiceScope;
 	scope = ServiceScope.PROTOTYPE, service = AgentInstanceResource.class
 )
 public class AgentInstanceResourceImpl extends BaseAgentInstanceResourceImpl {
+
+	@Override
+	public void getAgentInstanceSubscribe(SseEventSink sseEventSink) {
+		if (!FeatureFlagManagerUtil.isEnabled(
+				contextCompany.getCompanyId(), "LPD-62272")) {
+
+			throw new UnsupportedOperationException();
+		}
+
+		SseUtil.initialize(_sse, sseEventSink);
+	}
+
+	@Override
+	public AgentInstance postAgentInstance(AgentInstance agentInstance)
+		throws Exception {
+
+		if (!FeatureFlagManagerUtil.isEnabled(
+				contextCompany.getCompanyId(), "LPD-62272")) {
+
+			throw new UnsupportedOperationException();
+		}
+
+		WorkflowDefinition workflowDefinition =
+			_workflowDefinitionManager.getLatestWorkflowDefinition(
+				contextCompany.getCompanyId(), agentInstance.getType());
+
+		Map<String, Serializable> workflowContext =
+			WorkflowContextUtil.toWorkflowContext(
+				agentInstance.getContext(), contextHttpServletRequest,
+				agentInstance.getSseEventSinkKey());
+
+		workflowContext.put(
+			"accessToken",
+			contextHttpServletRequest.getHeader("Authorization"));
+		workflowContext.put("outBoundEventName", agentInstance.getType());
+		workflowContext.put(
+			"userToken",
+			contextHttpServletRequest.getHeader(
+				"Liferay-AI-Hub-Cell-On-Behalf-Of"));
+
+		WorkflowInstance workflowInstance =
+			_workflowInstanceManager.startWorkflowInstance(
+				contextCompany.getCompanyId(),
+				AccountEntryUtil.getUserAccountEntryGroupId(
+					contextUser.getUserId()),
+				contextUser.getUserId(), agentInstance.getType(),
+				workflowDefinition.getVersion(), null, workflowContext);
+
+		return new AgentInstance() {
+			{
+				setExternalReferenceCode(
+					() -> String.valueOf(
+						workflowInstance.getWorkflowInstanceId()));
+				setType(agentInstance::getType);
+			}
+		};
+	}
+
+	@Context
+	private Sse _sse;
+
+	@Reference
+	private WorkflowDefinitionManager _workflowDefinitionManager;
+
+	@Reference
+	private WorkflowInstanceManager _workflowInstanceManager;
+
 }
